web: fold What's new into the Help modal
Closes #481.

Removes the standalone "What's new" header chip and surfaces release
notes as a new section inside the Help modal, fetched lazily from
`GET /api/v1/changelog` on mount. The unseen-release dot (and its
`lastSeenChangelogVersion` zustand wiring) moves with it to the Help
button so returning visitors still see a "new release available"
affordance without a dedicated chip.

Coverage migrates from `WhatsNewModal.test.tsx` into a new
`HelpModal.test.tsx`; the `a11y.test.tsx` scans pick up a stubbed
`fetchChangelog` so axe still passes against the consolidated trigger
and open modal. While here, fix a pre-existing duplicate-key warning
inherited from `WhatsNewModal` — a release can have multiple sections
with the same name (the CHANGELOG splits notes by surface, so a
single release legitimately has multiple `### Added` blocks), so the
section key includes its position rather than only the name.
