+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Versions follow [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Web: **Header decluttered — What's new folded into the Help modal** ([#481](https://github.com/mgzwarrior/mgz-pkmn/issues/481)). Removes the standalone "What's new" header chip and surfaces the same release-notes content as a new section inside the Help modal, fetched lazily from `GET /api/v1/changelog` on mount. The unseen-release dot (and its `lastSeenChangelogVersion` zustand wiring) now lives on the Help button — returning visitors still see a "new release available" affordance without a dedicated chip competing with the brand on narrow viewports. Test coverage migrates from `WhatsNewModal.test.tsx` into a new `HelpModal.test.tsx`; the `a11y.test.tsx` scans continue to pass against the consolidated trigger and open modal.
+
 - Web: **Auth unification — Account panel** ([#491](https://github.com/mgzwarrior/mgz-pkmn/issues/491), slice 3 of 3). Closes the issue by giving the user something to look at: a new "Account" item in the signed-in chip's dropdown opens a Radix dialog listing each provider already attached to the row (with the per-identity email + a Disconnect button) and surfaces the missing providers as Connect actions. Connect GitHub / Google form-POST to the slice-2 `POST /api/v1/auth/link/{github,google}/start` endpoints — the session cookie travels with the same-origin POST, and the OAuth state from Authlib provides CSRF. Connect email expands inline into a magic-link form that hits `POST /api/v1/auth/link/magic/start` and renders the same "check your inbox" confirmation the sign-in dialog uses. Disconnect calls `DELETE /api/v1/auth/identities/{id}` and is disabled when only one identity remains so the last-provider safeguard is visible up-front rather than discovered as a 400. The Account panel also auto-opens when the SPA loads on `/account`, the path the link callback redirects to, and surfaces the slice-2 `?link_error=identity_already_linked&provider=…` query param as an inline alert ("That Google account is already linked to another mgz-pkmn account"). Pinned by 7 new tests in `web/src/components/AccountPanel.test.tsx` (identity rendering, disconnect call + last-identity disable, OAuth form-POST targets, magic-link submit + confirmation, 409 conflict surface, no spurious alert for unrelated query strings) plus 1 new `SignInChip.test.tsx` test covering the Account menu item.
 
 - API: **Auth unification — link / unlink endpoints** ([#491](https://github.com/mgzwarrior/mgz-pkmn/issues/491), slice 2 of 3). Adds the backend management surface on top of the `user_identities` foundation from slice 1: `POST /api/v1/auth/link/github/start` and `POST /api/v1/auth/link/google/start` stash a signed-in user's pending link marker in the session and redirect to the provider; `GET /api/v1/auth/link/github/callback` and `GET /api/v1/auth/link/google/callback` consume that marker, verify the OAuth callback, and attach the proved provider identity to the current `users` row even when the provider email differs from `users.email`; `POST /api/v1/auth/link/magic/start` sends a magic-link email pointed at `GET /api/v1/auth/link/magic/callback` for attaching an additional email identity. If the proved identity already belongs to another user, callbacks return `409` with an `identity_already_linked` detail body. `DELETE /api/v1/auth/identities/{id}` removes a linked provider for the signed-in user but rejects the last identity with `400 cannot_unlink_last_identity`. `GET /api/v1/me` now includes the current user's linked identities so slice 3's Account panel can render the list without an extra request. Pinned by `tests/test_auth_links.py` covering signed-in gating, OAuth and magic-link linking, conflict handling, and last-provider unlink protection.
