+++ web/src/components/HelpModal.tsx
@@ -1,13 +1,34 @@
 /**
  * HelpModal — onboarding/help dialog with sections covering queries,
- * settings, exports, and shortcuts. Includes a "Take the tour" button
- * that hands control off to the Tour component.
+ * settings, exports, shortcuts, and what's new. Includes a "Take the
+ * tour" button that hands control off to the Tour component.
+ *
+ * Also owns the "unseen release" affordance: a small dot on the trigger
+ * when the latest changelog version differs from the user's last-seen
+ * version (persisted in zustand). Opening the modal marks the latest
+ * version seen.
  */
 import * as Dialog from '@radix-ui/react-dialog'
 import { CircleHelp, X } from 'lucide-react'
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
+import { fetchChangelog } from '../api/client'
+import { useAppStore } from '../store'
+import type { ChangelogRelease } from '../types'
+import { renderInlineMarkdown } from '../utils/inlineMarkdown'
 
 const FIRST_VISIT_KEY = 'mgz-pkmn:seen-help'
+const REPO_CHANGELOG_URL =
+  'https://github.com/mgzwarrior/mgz-pkmn/blob/main/CHANGELOG.md'
+
+// Section-name → badge accent. Unknown names fall back to a neutral chip.
+const SECTION_ACCENT: Record<string, string> = {
+  Added: 'text-palm-600 border-palm-500/30 dark:text-palm-200 dark:border-palm-300/30',
+  Changed: 'text-sky-500 border-sky-400/30 dark:text-sky-300 dark:border-sky-400/30',
+  Fixed: 'text-sun-600 border-sun-400/30 dark:text-sun-300 dark:border-sun-400/30',
+  Removed: 'text-ember-500 border-ember-400/30 dark:text-ember-300 dark:border-ember-400/30',
+  Deprecated: 'text-sun-700 border-sun-500/30 dark:text-sun-400 dark:border-sun-500/30',
+  Security: 'text-ember-600 border-ember-500/30 dark:text-ember-300 dark:border-ember-500/30',
+}
 
 // localStorage access can throw in some browsers/contexts (Safari private
 // mode, embedded webviews with storage disabled, quota exhaustion). Wrap so
@@ -30,6 +51,18 @@ function markSeenHelp(): void {
   }
 }
 
+function formatDate(iso: string | null): string {
+  if (!iso) return ''
+  const d = new Date(`${iso}T00:00:00Z`)
+  if (Number.isNaN(d.getTime())) return iso
+  return d.toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+    timeZone: 'UTC',
+  })
+}
+
 interface Props {
   onStartTour: () => void
 }
@@ -37,6 +70,38 @@ interface Props {
 export function HelpModal({ onStartTour }: Props) {
   const [open, setOpen] = useState(false)
   const [hint, setHint] = useState(() => !readSeenHelp())
+  const [releases, setReleases] = useState<ChangelogRelease[] | null>(null)
+  const [releasesError, setReleasesError] = useState(false)
+
+  const lastSeen = useAppStore((s) => s.lastSeenChangelogVersion)
+  const setLastSeen = useAppStore((s) => s.setLastSeenChangelogVersion)
+
+  const latest = releases?.[0]?.version ?? null
+
+  // Fetch once on mount so the unseen-release dot can light up before the
+  // user ever opens the modal. A failure leaves `releases` null + the
+  // What's new section shows a fallback.
+  useEffect(() => {
+    let cancelled = false
+    fetchChangelog(5)
+      .then((data) => {
+        if (!cancelled) setReleases(data)
+      })
+      .catch(() => {
+        if (!cancelled) setReleasesError(true)
+      })
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  // Silent first-time catch-up: a visitor with no stored version is pinned
+  // to the current latest so the dot never shows on a first visit.
+  useEffect(() => {
+    if (latest && lastSeen === null) setLastSeen(latest)
+  }, [latest, lastSeen, setLastSeen])
+
+  const hasUnseen = !!latest && lastSeen !== null && latest !== lastSeen
 
   function handleOpenChange(next: boolean) {
     setOpen(next)
@@ -44,6 +109,8 @@ export function HelpModal({ onStartTour }: Props) {
       setHint(false)
       markSeenHelp()
     }
+    // Mark the latest version seen the moment the modal opens.
+    if (next && latest) setLastSeen(latest)
   }
 
   function handleTakeTour() {
@@ -59,10 +126,16 @@ export function HelpModal({ onStartTour }: Props) {
             hint ? 'ring-2 ring-palm-400 dark:ring-sun-300 animate-pulse' : ''
           }`}
           title="Help"
-          aria-label="Help"
+          aria-label={hasUnseen ? 'Help (new release available)' : 'Help'}
         >
           <CircleHelp size={15} />
           <span className="hidden sm:inline">Help</span>
+          {hasUnseen && (
+            <span
+              aria-hidden="true"
+              className="absolute -right-1 -top-1 h-2.5 w-2.5 rounded-full bg-palm-400 dark:bg-sun-300 ring-2 ring-sand-50 dark:ring-husk-400"
+            />
+          )}
         </button>
       </Dialog.Trigger>
 
@@ -152,6 +225,86 @@ export function HelpModal({ onStartTour }: Props) {
                   [<Kbd key="brand">Brand × 5</Kbd>, 'Click the logo five times for a surprise'],
                 ]}
               />
+            </Section>
+
+            <Section title="What's new">
+              {releasesError || (releases && releases.length === 0) ? (
+                <p className="text-coconut-400 dark:text-sand-300">
+                  Release notes couldn't be loaded right now. See the{' '}
+                  <a
+                    href={REPO_CHANGELOG_URL}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-palm-500 dark:text-sun-300 underline hover:text-palm-400 dark:hover:text-sun-200"
+                  >
+                    full changelog
+                  </a>
+                  .
+                </p>
+              ) : releases === null ? (
+                <p className="text-coconut-400 dark:text-sand-400">Loading release notes…</p>
+              ) : (
+                <>
+                  <ol className="space-y-6">
+                    {releases.map((release) => (
+                      <li
+                        key={release.version}
+                        className="border-l border-sand-200 dark:border-husk-100 pl-4"
+                      >
+                        <div className="flex flex-wrap items-baseline gap-x-3 gap-y-1">
+                          <h4 className="text-sm font-semibold text-coconut-700 dark:text-sand-50">
+                            v{release.version}
+                          </h4>
+                          {release.date && (
+                            <time className="text-xs text-coconut-400 dark:text-sand-400">
+                              {formatDate(release.date)}
+                            </time>
+                          )}
+                        </div>
+                        <div className="mt-2 space-y-3">
+                          {release.sections.map((section, sectionIdx) => (
+                            // A single release can have multiple sections with
+                            // the same name (e.g. two `### Added` blocks in
+                            // CHANGELOG.md), so index-by-position rather than
+                            // by section name to keep React keys unique.
+                            <div key={`${section.name}-${sectionIdx}`}>
+                              <span
+                                className={`inline-block rounded-full border px-2 py-0.5 text-xs font-medium ${
+                                  SECTION_ACCENT[section.name] ??
+                                  'text-coconut-400 dark:text-sand-300 border-sand-300 dark:border-husk-50'
+                                }`}
+                              >
+                                {section.name}
+                              </span>
+                              <ul className="mt-1.5 space-y-1.5">
+                                {section.entries.map((entry, i) => (
+                                  <li key={i} className="flex gap-2 leading-relaxed">
+                                    <span
+                                      aria-hidden="true"
+                                      className="mt-2 h-1 w-1 shrink-0 rounded-full bg-sand-300 dark:bg-husk-50"
+                                    />
+                                    <span>{renderInlineMarkdown(entry)}</span>
+                                  </li>
+                                ))}
+                              </ul>
+                            </div>
+                          ))}
+                        </div>
+                      </li>
+                    ))}
+                  </ol>
+                  <p className="mt-4 text-xs">
+                    <a
+                      href={REPO_CHANGELOG_URL}
+                      target="_blank"
+                      rel="noopener noreferrer"
+                      className="font-medium text-palm-500 dark:text-sun-300 hover:text-palm-400 dark:hover:text-sun-200 transition-colors"
+                    >
+                      Full changelog →
+                    </a>
+                  </p>
+                </>
+              )}
             </Section>
           </div>
